NetworkVariableProperty correctly alerts NetworkVariable to changes
NetworkVariableProperty now uses the NetworkVariables normal Value.Set when a new value is assigned in the inspector. This means changes made in the inspector use the same code path as changes made directly by scripts and ensures that changes made in the inspector are propogated out into the network.

Fixed the IsReadonly check in some of the reference property drawers. They were checking AssetManager when they shouldn't be.

Added ToString() methods for NetworkReferenceXX structs.