reworked reference types
They now store the reference internally instead of the id numbers. conversion is done at the time of de/serialization instead of at the time of access. Also reworked property drawers to update them to the new method and simplify them a bit. Now allows values to be assigned to reference types when not running.